refactor(search applications): rename engine api logic file
